requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==5.0
django-admin-persian-fonts==0.2
django-jalali-date==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
inflection==0.5.1
install==1.3.5
jdatetime==4.1.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rest-framework-simplejwt==0.0.2
rpds-py==0.15.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
